+++ components/engine_rocks/src/import.rs
@@ -62,57 +62,3 @@ impl IngestExternalFileOptions for RocksIngestExternalFileOptions {
         self.0.move_files(f);
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::cf_options::RocksColumnFamilyOptions;
-    use crate::db_options::{RocksDBOptions, RocksTitanDBOptions};
-    use crate::engine::RocksEngine;
-    use crate::import::RocksIngestExternalFileOptions;
-    use crate::sst::RocksSstWriterBuilder;
-    use crate::util::{new_engine, RocksCFOptions};
-    use engine_traits::CFHandleExt;
-    use engine_traits::CfName;
-    use engine_traits::ColumnFamilyOptions;
-    use engine_traits::ImportExt;
-    use engine_traits::IngestExternalFileOptions;
-    use engine_traits::Peekable;
-    use engine_traits::{DBOptions, TitanDBOptions};
-    use engine_traits::{SstWriter, SstWriterBuilder};
-    use std::fs;
-    use std::path::Path;
-    use tempfile::Builder;
-    use tikv_util::file::calc_crc32;
-
-    #[cfg(target_os = "linux")]
-    fn check_hard_link<P: AsRef<Path>>(path: P, nlink: u64) {
-        use std::os::linux::fs::MetadataExt;
-        assert_eq!(fs::metadata(path).unwrap().st_nlink(), nlink);
-    }
-
-    #[cfg(not(target_os = "linux"))]
-    fn check_hard_link<P: AsRef<Path>>(_: P, _: u64) {
-        // Just do nothing
-    }
-
-    fn check_db_with_kvs(db: &RocksEngine, cf: &str, kvs: &[(&str, &str)]) {
-        for &(k, v) in kvs {
-            assert_eq!(
-                db.get_value_cf(cf, k.as_bytes()).unwrap().unwrap(),
-                v.as_bytes()
-            );
-        }
-    }
-
-    fn gen_sst_with_kvs(db: &RocksEngine, cf: CfName, path: &str, kvs: &[(&str, &str)]) {
-        let mut writer = RocksSstWriterBuilder::new()
-            .set_db(db)
-            .set_cf(cf)
-            .build(path)
-            .unwrap();
-        for &(k, v) in kvs {
-            writer.put(k.as_bytes(), v.as_bytes()).unwrap();
-        }
-        writer.finish().unwrap();
-    }
-}
